+++ tests/unit/test_model.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from typing import Any, Dict
 from uuid import UUID
 
 import pytest
@@ -15,25 +16,25 @@ from restio.fields import (
     TupleField,
     TupleModelField,
 )
-from restio.model import MODEL_UPDATE_EVENT, BaseModel
+from restio.model import MODEL_UPDATE_EVENT, BaseModel, ModelMeta
 
 
 class ModelSinglePKInt(BaseModel):
-    id: IntField = IntField(pk=True)
+    id: IntField = IntField(pk=True, allow_none=True)
 
 
 class ModelSinglePKStr(BaseModel):
-    id: StrField = StrField(pk=True)
+    id: StrField = StrField(pk=True, allow_none=True)
 
 
 class ModelDoublePKIntStr(BaseModel):
-    id: IntField = IntField(pk=True)
-    key: StrField = StrField(pk=True)
+    id: IntField = IntField(pk=True, allow_none=True)
+    key: StrField = StrField(pk=True, allow_none=True)
 
 
 class ModelDoublePKStrInt(BaseModel):
-    key: StrField = StrField(pk=True)
-    id: IntField = IntField(pk=True)
+    key: StrField = StrField(pk=True, allow_none=True)
+    id: IntField = IntField(pk=True, allow_none=True)
 
 
 class TestModelPrimaryKeys:
@@ -82,29 +83,35 @@ class TestModelPrimaryKeys:
 
 
 class ModelA(BaseModel):
-    id: IntField = IntField(pk=True)
-    a: IntField = IntField()
-    b: StrField = StrField()
-    c: TupleField = TupleField(str)
-    d: FrozenSetField = FrozenSetField(int)
-    e: BoolField = BoolField()
+    id: IntField = IntField(default=0, pk=True)
+    a: IntField = IntField(default=0)
+    b: StrField = StrField(default="")
+    c: TupleField = TupleField(str, default_factory=tuple)
+    d: FrozenSetField = FrozenSetField(int, default_factory=frozenset)
+    e: BoolField = BoolField(default=False)
 
 
 class ModelB(BaseModel):
     ref: ModelField[ModelA] = ModelField(ModelA)
-    c: StrField = StrField()
+    c: StrField = StrField(default="")
 
 
 class ModelC(BaseModel):
-    ref_tuple: TupleModelField[ModelB] = TupleModelField(ModelB)
-    ref_frozenset: FrozenSetModelField[ModelB] = FrozenSetModelField(ModelB)
-    d: StrField = StrField()
+    ref_tuple: TupleModelField[ModelB] = TupleModelField(ModelB, default_factory=tuple)
+    ref_frozenset: FrozenSetModelField[ModelB] = FrozenSetModelField(
+        ModelB, default_factory=frozenset
+    )
+    d: StrField = StrField(default="")
+
+
+default_model = ModelA()
 
 
 class TestModel:
     def test_model_internal_uuid(self):
         class Model(BaseModel):
-            pass
+            def __init__(self):
+                assert not self._internal_id
 
         model_a = Model()
         model_b = Model()
@@ -114,6 +121,17 @@ class TestModel:
         assert isinstance(model_a._internal_id, UUID)
         assert isinstance(model_b._internal_id, UUID)
         assert model_a._internal_id != model_b._internal_id
+
+    def test_model_hash(self):
+        class Model(BaseModel):
+            def __init__(self):
+                assert not self._hash
+
+        model_a = Model()
+
+        assert model_a
+        assert isinstance(model_a._hash, int)
+        assert hash(model_a) == model_a._hash
 
     def test_model_persistent_values(self):
         class Model(BaseModel):
@@ -157,11 +175,122 @@ class TestModel:
 
         assert model_a != model_b
 
+    def test_model_meta_defaults(self):
+        class Model(BaseModel):
+            class Meta:
+                fields = {"a", "b"}
+                primary_keys = {"a"}
+
+        model = Model()
+
+        assert model._meta.init
+        assert model._meta.init_ignore_extra
+        assert model._meta.fields == {}
+        assert model._meta.primary_keys == {}
+
+    def _assert_model_meta(self, model: BaseModel, meta_dict: Dict[str, Any]):
+        meta = ModelMeta()
+
+        for meta_attr in meta.__slots__:
+            # assert child class
+            if meta_attr in meta_dict:
+                assert getattr(model._meta, meta_attr) == meta_dict[meta_attr]
+            else:
+                default = getattr(meta, meta_attr)
+                assert getattr(model._meta, meta_attr) == default
+
+    @pytest.mark.parametrize(
+        "meta_field, value", [("init", False), ("init_ignore_extra", False)]
+    )
+    def test_model_meta_configurable(self, meta_field, value):
+        meta_dict = {meta_field: value}
+
+        class Model(BaseModel):
+            Meta = type("Meta", (), meta_dict)
+
+        model = Model()
+
+        self._assert_model_meta(model, meta_dict)
+
+    @pytest.mark.parametrize(
+        "meta_field, value_parent, value_child",
+        [("init", False, True), ("init_ignore_extra", False, True)],
+    )
+    def test_model_meta_configurable_inheritance_single_substitution(
+        self, meta_field, value_parent, value_child
+    ):
+        meta_child_dict = {meta_field: value_child}
+        meta_parent_dict = {meta_field: value_parent}
+
+        class ParentModel(BaseModel):
+            Meta = type("Meta", (), meta_parent_dict)
+
+        class ChildModel(ParentModel):
+            Meta = type("Meta", (), meta_child_dict)
+
+        parent_model = ParentModel()
+        child_model = ChildModel()
+
+        self._assert_model_meta(parent_model, meta_parent_dict)
+        self._assert_model_meta(child_model, meta_child_dict)
+
+    def test_model_meta_configurable_inheritance_multiple_substitution(self):
+        class ParentModel(BaseModel):
+            class Meta:
+                init = False
+
+        class ChildModel(ParentModel):
+            class Meta:
+                init = True
+                init_extra_ignore = False
+
+        class GrandChildModel(ChildModel):
+            class Meta:
+                init = False
+                init_extra_ignore = True
+
+        parent_model = ParentModel()
+        child_model = ChildModel()
+        grand_child_model = GrandChildModel()
+
+        self._assert_model_meta(parent_model, {"init": False})
+        self._assert_model_meta(child_model, {"init": True, "init_extra_ignore": False})
+        self._assert_model_meta(
+            grand_child_model, {"init": False, "init_extra_ignore": True}
+        )
+
+    def test_all_field_types_in_model(self):
+        class AllFieldsModel(BaseModel):
+            int_key = IntField(pk=True)
+            str_key = StrField(pk=True)
+            int_field = IntField()
+            str_field = StrField()
+            bool_field = BoolField()
+            tuple_field = TupleField(str)
+            set_field = FrozenSetField(int)
+            model_field = ModelField(ModelA)
+            tuple_model_field = TupleModelField(ModelA)
+            set_model_field = FrozenSetModelField(ModelA)
+            non_field: str
+
+        assert AllFieldsModel._meta.fields.keys() == {
+            "int_key",
+            "str_key",
+            "int_field",
+            "str_field",
+            "bool_field",
+            "tuple_field",
+            "set_field",
+            "model_field",
+            "tuple_model_field",
+            "set_model_field",
+        }
+
     @pytest.mark.parametrize(
         "model, expected_fields, expected_dependency_fields",
         [
             (
-                ModelA(),
+                default_model,
                 {"id": 0, "a": 0, "b": "", "c": tuple(), "d": frozenset(), "e": False},
                 {},
             ),
@@ -178,23 +307,211 @@ class TestModel:
         assert model.fields == expected_fields
         assert model.dependency_fields == expected_dependency_fields
 
+    @pytest.mark.parametrize(
+        "field_type, default",
+        [
+            (IntField, 1),
+            (StrField, "a"),
+            (BoolField, True),
+            (lambda default: TupleField(int, default_factory=default), lambda: (1, 2)),
+            (
+                lambda default: FrozenSetField(str, default_factory=default),
+                lambda: frozenset(["a", "b"]),
+            ),
+            (lambda default: ModelField(ModelA, default=default), default_model),
+            (
+                lambda default: TupleModelField(ModelA, default_factory=default),
+                lambda: (default_model,),
+            ),
+            (
+                lambda default: FrozenSetModelField(ModelA, default_factory=default),
+                lambda: frozenset({default_model}),
+            ),
+        ],
+    )
+    def test_model_fields_custom_default(self, field_type, default):
+        class Model(BaseModel):
+            field = field_type(default=default)
+
+        obj = Model()
+
+        assert Model._meta.fields.keys() == {"field"}
+        assert obj.field == default() if callable(default) else default
+
     def test_model_fields_inheritance(self):
         class Model(BaseModel):
             a: IntField = IntField(default=1)
             non_field_a: int = 0
 
         class ChildModel(Model):
-            b: StrField = StrField()
+            b: StrField = StrField(default="v")
             non_field_b: str = ""
+
+        class GrandChildModel(ChildModel):
+            b: IntField = IntField(default=2)
 
         model = Model()
         child_model = ChildModel()
+        grand_child_model = GrandChildModel()
 
         assert Model._meta.fields.keys() == {"a"}
         assert ChildModel._meta.fields.keys() == {"a", "b"}
+        assert GrandChildModel._meta.fields.keys() == {"a", "b"}
 
         assert model.fields == {"a": 1}
-        assert child_model.fields == {"a": 1, "b": ""}
+        assert child_model.fields == {"a": 1, "b": "v"}
+        assert grand_child_model.fields == {"a": 1, "b": 2}
+
+    def test_model_generated_constructor(self):
+        class Model(BaseModel):
+            a: IntField = IntField(default=1)
+            b: StrField = StrField(default="default")
+
+        model = Model()
+        model_const = Model(a=2, b="default_const")
+
+        assert model.a == 1
+        assert model.b == "default"
+
+        assert model_const.a == 2
+        assert model_const.b == "default_const"
+
+    def test_model_generated_constructor(self):
+        class Model(BaseModel):
+            a: IntField = IntField(default=1)
+            b: StrField = StrField(default="default")
+
+        model = Model()
+        model_const = Model(a=2, b="default_const")
+        model_part_const = Model(a=3, c="ignored")
+
+        assert model.a == 1
+        assert model.b == "default"
+
+        assert model_const.a == 2
+        assert model_const.b == "default_const"
+
+        assert model_part_const.a == 3
+        assert model_part_const.b == "default"
+
+    def test_model_generated_constructor_partially_default(self):
+        class Model(BaseModel):
+            a: IntField = IntField(default=1)
+            b: StrField = StrField()
+
+        model = Model(b="value")
+
+        assert model.a == 1
+        assert model.b == "value"
+
+    def test_model_non_generated_constructor(self):
+        class Model(BaseModel):
+            class Meta:
+                init = False
+
+            a: IntField = IntField(default=1)
+            b: StrField = StrField(default="default")
+
+        model = Model(a=2, b="default_const")
+
+        assert model.a == 1
+        assert model.b == "default"
+
+    def test_model_generated_constructor_ignored_extras(self):
+        class Model(BaseModel):
+            a: IntField = IntField()
+            b: StrField = StrField()
+
+        Model(a=2, b="default_const", c="not_ignored")
+
+    def test_model_generated_constructor_not_ignored_extras(self):
+        class Model(BaseModel):
+            class Meta:
+                init_ignore_extra = False
+
+            a: IntField = IntField()
+            b: StrField = StrField()
+
+        with pytest.raises(ValueError, match="Invalid argument"):
+            Model(a=2, b="default_const", c="not_ignored")
+
+    def test_model_generated_constructor_missing_values(self):
+        class Model(BaseModel):
+            a: IntField = IntField()
+            b: StrField = StrField()
+
+        with pytest.raises(ValueError, match="Can't initialize field"):
+            Model()
+
+        with pytest.raises(ValueError, match="Can't initialize field"):
+            Model(a=2)
+
+        with pytest.raises(ValueError, match="Can't initialize field"):
+            Model(b="")
+
+    def test_model_generated_constructor_partially_initialized(self):
+        class Model(BaseModel):
+            a: IntField = IntField()
+            b: StrField = StrField(init=False, default="default")
+
+        model = Model(a=1)
+        model_part = Model(a=1, b="non-default")
+
+        assert model.a == 1
+        assert model.b == "default"
+
+        assert model_part.a == 1
+        assert model_part.b == "default"
+
+    def test_model_generated_constructor_not_initialized_without_default(self):
+        class Model(BaseModel):
+            a: IntField = IntField()
+            b: StrField = StrField(init=False)
+
+        with pytest.raises(ValueError, match="Can't initialize field"):
+            Model(a=1)
+
+        with pytest.raises(ValueError, match="Can't initialize field"):
+            Model(a=1, b="")
+
+    def test_model_generated_constructor_ignored_extras_not_initialized(self):
+        class Model(BaseModel):
+            class Meta:
+                init_ignore_extra = False
+
+            a: IntField = IntField()
+            b: StrField = StrField(init=False)
+
+        with pytest.raises(ValueError, match="cannot be initialized"):
+            Model(a=2, b="default_const")
+
+    def test_model_manual_constructor_with_super_call(self):
+        class Model(BaseModel):
+            a: IntField = IntField()
+            b: StrField = StrField()
+
+            def __init__(self, **kwargs):
+                super().__init__(**kwargs)
+                self.a = self.a + 1
+
+        model = Model(a=1, b="val")
+
+        assert model.a == 2
+        assert model.b == "val"
+
+    def test_model_manual_constructor_without_super_call(self):
+        class Model(BaseModel):
+            a: IntField = IntField()
+            b: StrField = StrField()
+
+            def __init__(self, a: int, b: str):
+                self.a = a + 1
+                self.b = b + "-added"
+
+        model = Model(a=1, b="val")
+
+        assert model.a == 2
+        assert model.b == "val-added"
 
     def test_model_update_persistent_values(self):
         model = ModelA()